#50 fix warnings

+++ src/ChaiVM/interpreter/reg-file.cpp
@@ -5,11 +5,11 @@
 
 namespace chai::interpreter {
 
-chai::chsize_t &RegisterFile::operator[](int n) & {
+chai::chsize_t &RegisterFile::operator[](size_t n) & {
     assert(n <= Size);
     return registers_[n];
 }
-const chsize_t &RegisterFile::operator[](int n) const & {
+const chsize_t &RegisterFile::operator[](size_t n) const & {
     assert(n <= Size);
     return registers_[n];
 }
@@ -22,7 +22,7 @@ RegisterFile::RegisterFile(chsize_t pc) : acc_(0), pc_(pc), registers_{} {}
 void RegisterFile::dump() {
     std::cout << "pc = " << pc_ << ", acc = " << std::bit_cast<int64_t>(acc_)
               << " = " << std::bit_cast<double>(acc_) << std::endl;
-    for (int i = 0; i < Size; ++i) {
+    for (size_t i = 0; i < Size; ++i) {
         if (registers_[i] != 0) {
             std::cout << "rf[" << i
                       << "] = " << std::bit_cast<int64_t>(registers_[i])
